Move ExecutionNode state back to inner class
