+++ engine/runtime/src/main/java/org/enso/interpreter/instrument/ReplDebuggerInstrument.java
@@ -207,6 +207,37 @@ public class ReplDebuggerInstrument extends TruffleInstrument {
         exit();
       }
     }
+
+    /**
+     * State of the execution node.
+     *
+     * As the execution nodes are reused by Truffle, the nested nodes share
+     * state. If execution of a nested node fails, to ensure consistent state of
+     * the parent node, its state has to be restored.
+     */
+    private static class ReplExecutionEventNodeState {
+      private final Object lastReturn;
+      private final Object lastState;
+      private final CallerInfo lastScope;
+
+      ReplExecutionEventNodeState(Object lastReturn, Object lastState, CallerInfo lastScope) {
+        this.lastReturn = lastReturn;
+        this.lastState = lastState;
+        this.lastScope = lastScope;
+      }
+
+      Object getLastReturn() {
+        return lastReturn;
+      }
+
+      Object getLastState() {
+        return lastState;
+      }
+
+      CallerInfo getLastScope() {
+        return lastScope;
+      }
+    }
   }
 
 }
